Запустили колесо с пульта

+++ Bastet/Inc/kinematica.h
@@ -8,7 +8,7 @@
 
 double sign(double a);
 void kinematica(uint32_t mode, Motor_t* Motors, servoTarget_t* Servo);
-void normaliz(uint32_t vel_mean, uint32_t rx_mean, uint32_t dir_mean, uint32_t ry_mean);
+void normaliz(uint16_t vel_mean, uint16_t rx_mean, uint16_t dir_mean, uint16_t ry_mean);
 
 #define pi 3.14159265
 #define deg 57.29577951
